Introducing coordinate_types.hpp!
Coordinate arithmetic is absolutely fraught with challenges.  In order to
reduce the possibility of bugs, we've introduced an extremely pedantic system
of coordinate types.  The busclique algorithms use two kinds of coordinate
systems:

  * matrix coordinates (y, x) which are featured in the chimera coordinate
    system (y, x, u, k) -- where u indicates the 'orientation' of a qubit,
    either vertical(u=0) or horizontal(u=1) and k indicates the 'shore index'
    of a qubit (irrelevant here)

  * relative coordinates (w, z) which are features in the pegasus and zephyr
    coordinate systems (u, w, k[, j], z) -- where u is the orientation as
    above and k and j are similarly irrelevant.

The relationship between matrix and relative coordinates is that

  * (y, x, u=0) is equivalent to (z=y, w=x, u=0), and

  * (y, x, u=1) is equivalent to (w=y, z=x, u=1)

We have encoded this equivalence in four types, size_y, size_x, size_w, and
size_z.  Each type supports arithmetic with itself, but forbids confusion.  For
example,

    size_y y0, y1;
    size_x x0, x0;
    (y0 + y1*y0 + y1)/(y1-y0) // cool
    y0 + x0;                  // compile error!

But, we need to convert between matrix and relative coordinates frequently!
We do that with the `vert` and `horz` methods:

    bool u = 0;
    size_y y;
    size_x x;
    size_w w0, w1;
    size_z z0, z1;
    w1 = w0 + y.vert();
    z1 = w0 + x.vert();

Sometimes, we *really* need to mix things up.  There are two ways to convert
a coordinate to a size_t: implicitly through `coordinate_base::operator size_t`,
or explicitly through `coordinate_base::index()`.  The implicit conversion is
private, to prevent accidental implicit conversions.  The `coordinate_converter`
class is a friend to the coordinate types, and that's the only place we need to
be extremely careful.  The `index()` method is used in very few one-off
instances.  Ideally, it should only be used for indexing into arrays:

    size_y y;
    uint8_t mem[100];
    mem[y.index()];

This entire file, and refactor, is a very heavy-handed approach to fixing a bug
(segfaulting non-square Chimera graphs).  However, it reduces a significant
amount of anxiety in developing and extending this library.  In small tests,
code involving coordinate classes is optimized down to identical assembly to the
same code using size_t in their place.  We'll see if that's true in a complex
codebase...